pass `(server_id, buffer)` to LspAdapter::process_diagnostics
converting diagnostics between Zed and lsp seems inefficient...
